# docs/swagger.yaml
definitions:
...
  contribution.ArticleReq:
    properties:
      description:
        type: string
      link:
        type: string
      title:
        type: string
    type: object
  contribution.ContributionCreateReq:
    properties:
      article:
        $ref: '#/definitions/contribution.ArticleReq'
      images:
        items:
          $ref: '#/definitions/contribution.ImageCreateReq'
        type: array
    type: object
  contribution.ContributionRes:
    properties:
      articleId:
        type: integer
      contributeSessionId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/contribution.UserRes'
    type: object
  contribution.ContributionStatusReq:
    properties:
      status:
        type: string
    type: object
  contribution.ContributionUpdateReq:
    properties:
      article:
        $ref: '#/definitions/contribution.ArticleReq'
      images:
        items:
          $ref: '#/definitions/contribution.ImageCreateReq'
        type: array
    type: object
  contribution.ImageCreateReq:
    properties:
      key:
        type: string
      title:
        type: string
    type: object
  contribution.ImageRes:
    properties:
      key:
        type: string
      link:
        type: string
      title:
        type: string
    type: object
  contribution.PaginateComposition:
    properties:
      currentPage:
        type: integer
      data:
        items:
          $ref: '#/definitions/contribution.ContributionRes'
        type: array
      lastPage:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
  contribution.UserRes:
    properties:
      email:
        type: string
      facultyId:
        type: integer
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  faculty.FacultyCreateReq:
    properties:
      name:
        type: string
    type: object
  faculty.FacultyResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  faculty.FacultyUpdateReq:
    properties:
      name:
        type: string
    type: object
  faculty.PaginateComposition:
    properties:
      currentPage:
        type: integer
      data:
        items:
          $ref: '#/definitions/faculty.FacultyResponse'
        type: array
      lastPage:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
...
  user.PaginateComposition:
    properties:
      currentPage:
        type: integer
      data:
        items:
          $ref: '#/definitions/user.UserResponse'
        type: array
      lastPage:
        type: integer
      perPage:
        type: integer
      total:
        type: integer
    type: object
  user.UserCreateReq:
    properties:
      email:
        type: string
      facultyId:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        enum:
        - admin
        - marketing_manager
        - marketing_coordinator
        - student
        - guest
        type: string
    type: object
  user.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      facultyId:
        type: integer
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
  user.UserUpdateReq:
    properties:
      email:
        type: string
      facultyId:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  title: "123"
paths:
...
  /contributions:
    get:
      consumes:
      - application/json
      description: List contributions
      parameters:
      - in: query
        name: contributionSessionId
        type: integer
      - in: query
        name: facultyId
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - enum:
        - accepted
        - rejected
        - reviewing
        in: query
        name: status
        type: string
      - in: query
        name: studentId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contribution.PaginateComposition'
      security:
      - ApiKeyAuth: []
      summary: List contributions
      tags:
      - Contributions
    post:
      consumes:
      - application/json
      description: Create a contribution
      parameters:
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contribution.ContributionCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contribution.ContributionRes'
      security:
      - ApiKeyAuth: []
      summary: Create a contribution
      tags:
      - Contributions
  /contributions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a contribution
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a contribution
      tags:
      - Contributions
    get:
      consumes:
      - application/json
      description: get contribution by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contribution.ContributionRes'
      security:
      - ApiKeyAuth: []
      summary: Show a contribution
      tags:
      - Contributions
    put:
      consumes:
      - application/json
      description: Update a contribution
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/contribution.ContributionUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/contribution.ContributionRes'
      security:
      - ApiKeyAuth: []
      summary: Update a contribution
      tags:
      - Contributions
  /contributions/{id}/images:
    get:
      consumes:
      - application/json
      description: Get contribution images
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/contribution.ImageRes'
            type: array
      security:
        - ApiKeyAuth: [ ]
      summary: Get contribution images
      tags:
        - Contributions
  /contributions/{id}/status:
    post:
      consumes:
        - application/json
      description: Update contribution status
      parameters:
        - description: ID
          in: path
          name: id
          required: true
          type: integer
        - description: update
          in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/contribution.ContributionStatusReq'
      produces:
        - application/json
      responses:
        "200":
          description: ""
      security:
        - ApiKeyAuth: [ ]
      summary: Update contribution status
      tags:
        - Contributions
  /faculties:
    get:
      consumes:
        - application/json
      description: List faculties
      parameters:
        - in: query
          name: limit
          type: integer
        - in: query
          name: page
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faculty.PaginateComposition'
      security:
      - ApiKeyAuth: []
      summary: List faculties
      tags:
      - Faculties
    post:
      consumes:
      - application/json
      description: Create a faculty
      parameters:
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/faculty.FacultyCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faculty.FacultyResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a faculty
      tags:
      - Faculties
  /faculties/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a faculty
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a faculty
      tags:
      - Faculties
    get:
      consumes:
      - application/json
      description: get faculty by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faculty.FacultyResponse'
      security:
      - ApiKeyAuth: []
      summary: Show a faculty
      tags:
      - Faculties
    put:
      consumes:
      - application/json
      description: Update a faculty
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/faculty.FacultyUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/faculty.FacultyResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a faculty
      tags:
      - Faculties
...
  /users:
    get:
      consumes:
      - application/json
      description: List users
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - enum:
        - admin
        - marketing_manager
        - marketing_coordinator
        - student
        - guest
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.PaginateComposition'
      security:
      - ApiKeyAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a user
      parameters:
      - description: create user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Show a user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a user
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"